+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data expansion feature to Stripe `checkoutSession` function.
 - Stripe checkout session creation `mode` to payload data.
 - Airtable module.
+- Order module.
 
 ### Changed
 - Replace checkout line items deprecated data parameters with `price_data`.